requirements.txt
streamlit
openai
langchain
pydantic
pytesseract
Pillow
python-dotenv